requirements.txt
taipy
openai
python-dotenv